Moving debug streamer to CommonUtils
