+++ Common/Utils/CMakeLists.txt
@@ -23,6 +23,7 @@ o2_add_library(CommonUtils
                        src/BoostHistogramUtils.cxx
                        src/NameConf.cxx
                        src/IRFrameSelector.cxx
+                       src/DebugStreamer.cxx
                PUBLIC_LINK_LIBRARIES ROOT::Hist ROOT::Tree Boost::iostreams O2::CommonDataFormat O2::Headers
                                      FairLogger::FairLogger)
 
@@ -45,7 +46,8 @@ o2_target_root_dictionary(CommonUtils
                                   include/CommonUtils/VerbosityConfig.h
                                   include/CommonUtils/FileFetcher.h
                                   include/CommonUtils/NameConf.h
-                                  include/CommonUtils/IRFrameSelector.h)
+                                  include/CommonUtils/IRFrameSelector.h
+                                  include/CommonUtils/DebugStreamer.h)
 
 o2_add_test(TreeStream
             COMPONENT_NAME CommonUtils
