+++ Common/Utils/src/CommonUtilsLinkDef.h
@@ -39,4 +39,7 @@
 
 #pragma link C++ class o2::utils::IRFrameSelector + ;
 
+#pragma link C++ struct o2::gpu::ParameterDebugStreamer + ;
+#pragma link C++ class o2::gpu::DebugStreamer + ;
+
 #endif
